feat: categBox  component
